Make time adjustment feature more efficient and correct, fixes #173
Replaced `RestUtils#getAWSTimeAdjustment` method with the improved
function `RestUtils#calculateTimeAdjustmentOffset` which uses the
'Date' header in a given response to calculate the time offset between
the local system and a service endpoint, rather than making an extra
HTTP connection to another server to lookup the correct time.

All code that used to rely on the obsolete and removed function has
been updated to use the new approach.

Also fixed a few incorrect Javadoc @throws claims, while I was at it.