Cache bucket-to-region mappings to assist AWS version 4 signing, re #183
Add a whole lot of region-aware smarts to JetS3t to try and hide the
fact the AWS signature version 4 requires much more fore-knowledge of
bucket-to-region mappings than is reasonable. All this work is verging
much more towards magic than I'm happy with and is sure to lead to
strange issues, but the alternative is dire for users.

`RestStorageService` now uses the new `RegionEndpointCache` to cache
known mappings from bucket names to regions. These mappings could be
inserted manually, and are also added automatically by JetS3t when it
is able to learn about bucket-to-region mappings (such as from prior
requests to the same bucket using a region-specific Host, or after
requests that fail with a clear region-related region where the correct
region is supplied by S3).

This all means that HEAD requests to a bucket in an unknown region can
succeed for the first time, provided the bucket has a mapping in
`RegionEndpointCache`.