Close HTTP response immediately when object input stream is closed, re #221 #182
When an object's data input stream is closed, close the response and
underlying connection immediately instead of consuming it, which was
causing ALL the object content to be downloaded.

This change follows the advice in section "1.1.5. Ensuring release of
low level resources" [1] which states:

   The difference between closing the content stream and closing the
   response is that the former will attempt to keep the underlying
   connection alive by consuming the entity content while the latter
   immediately shuts down and discards the connection.

WARNING: This change does not yet perform as expected, since JetS3t's
continued use of out-of-date HttpClient APIs for building and
configuring the client does not produce a CloseableHttpClient that
can be closed as required. For the time being, closing the response
just consumes the response entity (and retrieves all its data).
Fixing this will likely require updating our usage of deprecated
HttpClient APIs.

[1]: http://hc.apache.org/httpcomponents-client-4.5.x/tutorial/html/fundamentals.html#d5e145