Trim whitespace in metadata to comply with HTTP header conventions, re #230
Trim the string metadata value before uploading to a REST storage
service to ensure the value complies with HTTP header conventions which
expect that excess whitespace is removed.

This change ensures that the metadata values uploaded initially to S3
etc match the values that will be subsequently downloaded from the
service, at which point any excess whitespace uploaded to the service
gets stripped anyway.

The change may also help in situations where an intermediary proxy
modifies HTTP headers to be compliant by stripping the excess
whitespace, but in the process could break the request signature which
is only valid for the metadata value with the excess whitespace.